feat : 생활꿀팁, 꿀템 게시물 조회 API 구조 구현

+++ src/main/java/umc/demoday/whatisthis/domain/post/controller/PostController.java
@@ -20,7 +20,7 @@ public class PostController {
     @GetMapping("/{post-id}")
     @ResponseStatus(HttpStatus.ACCEPTED)
     @Operation(summary = "생활꿀팁 or 생활꿀팁 페이지 조회 API -by 천성호")
-    public CustomResponse<PostResponseDTO.GgulPostResponseDTO> getGgulPost(@PathVariable Long postId){
+    public CustomResponse<PostResponseDTO.GgulPostResponseDTO> getGgulPost(@PathVariable Integer postId){
 
         PostResponseDTO.GgulPostResponseDTO response = PostServiceImpl.getGgulPost(postId);
         return CustomResponse.onSuccess(response);
